First section of refactoring.
