+++ rpi/include/pathFinder.hpp
@@ -1,0 +1,10 @@
+#ifndef _PATH_FINDER_HPP_
+#define _PATH_FINDER_HPP_
+
+class PathFinder
+{
+
+};
+
+
+#endif
