--- rpi/pathfinder.h
@@ -1,12 +1,0 @@
-#ifndef PATH_FINDER_H
-#define PATH_FINDER_H
-
-class PathFinder
-{
-
-
-
-};
-
-
-#endif
