+++ rpi/src/pathNode.cpp
@@ -1,4 +1,4 @@
-#include "pathNode.h"
+#include "pathNode.hpp"
 
 PathNode::PathNode(int _x, int _y, int _direction, int _cost):x(_x), y(_y), direction(_direction), cost(_cost)
 {
